feat: add texture paths and color definitions to map files; implement map initialization and player parsing functions
