+++ map2.cub
@@ -1,3 +1,10 @@
+NO ./texture/wolfenstein/bluestone.xpm
+SO ./texture/wolfenstein/colorstone.xpm
+WE ./texture/wolfenstein/eagle.xpm
+EA ./texture/wolfenstein/redbrick.xpm
+F 220,100,0
+C 225,30,0
+
 1111111111
 110S000000
 1100000001
