+++ map3.cub
@@ -1,3 +1,10 @@
+NO ./texture/wolfenstein/bluestone.xpm
+SO ./texture/wolfenstein/colorstone.xpm
+WE ./texture/wolfenstein/eagle.xpm
+EA ./texture/wolfenstein/redbrick.xpm
+F 220,100,0
+C 225,30,0
+
 	 11111111111111111111111111111111111		 11111111111111111111111111111111111
 	 1S000000000000000000000000000000001		 10001
 	 10000000000000000000000000000000001	     11111
